accident-data-collection/scraper/requirements.txt
requests>=2.31
boto3>=1.34
pyarrow>=14.0
python-dateutil>=2.8
